.history/requirements_20250103093748.txt
chardet==5.2.0
opencv-python==4.10.0.84
pillow==11.0.0
dashscope==1.20.14
google-cloud-translate==3.18.0
torch==2.5.1
openai-whisper==20240930
pydub==0.25.1
requests==2.32.3
TTS==0.22.0
tinytag==2.0.0
gTTS==2.5.4
edge-tts==7.0.0
opencv-python==4.10.0.84